security: Started removing nodeIntegration.
We now have a SorusNative object exposed to the page from the preload script. Note that this currently has an issue that I couldn't fix (always updating even when there isn't an update).
