+++ index.html
@@ -235,7 +235,6 @@
     <div></div>
 
     <script src="./js/optionsmanager.js"></script>
-    <script src="./js/updateUtils.js"></script>
     <script src="./js/accounts.js"></script>
     <script src="./js/archive.js"></script>
     <script src="./js/sidebar.js"></script>
