+++ js/launch.js
@@ -72,32 +72,32 @@ async function extractAll() {
 }
 
 function checkAndDownloadSorus() {
-return new Promise((resolve, reject) => {
+  return new Promise((resolve, reject) => {
     let counter = 0;
     if (!fs.existsSync(userDataPath + '/mc/Sorus/client/Core.jar') || 
         !fs.existsSync(userDataPath + '/mc/Sorus/client/' + options.mc_ver + '.jar') || 
         !fs.existsSync(userDataPath + '/mc/Sorus/client/JavaAgent.jar') ||
-        checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/Core.jar", "Core") ||
-        checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/versions/" + options.mc_ver + ".jar", options.mc_ver) ||
-        checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/environments/JavaAgent.jar", "JavaAgent")) {
+        SorusNative.checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/Core.jar", "Core") ||
+        SorusNative.checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/versions/" + options.mc_ver + ".jar", options.mc_ver) ||
+        SorusNative.checkUpdate("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/environments/JavaAgent.jar", "JavaAgent")) {
       try {
         playbtn_status.innerHTML = "Downloading Core.jar"
         downloadSorus("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/Core.jar", "Core").then((dest) => {
-          console.log('Succesfully downloaded ' + dest);
+          console.log('Successfully downloaded ' + dest);
           counter++;
           counter > 2 && resolve();
         }).catch(console.error);
 
         playbtn_status.innerHTML = "Downloading Sorus " + options.mc_ver
         downloadSorus("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/versions/" + options.mc_ver + ".jar", options.mc_ver).then((dest) => {
-          console.log('Succesfully downloaded ' + dest);
+          console.log('Successfully downloaded ' + dest);
           counter++;
           counter > 2 && resolve();
         }).catch(console.error);
 
         playbtn_status.innerHTML = "Downloading JavaAgent.jar"
         downloadSorus("https://raw.githubusercontent.com/SorusClient/Sorus-Resources/master/client/environments/JavaAgent.jar", "JavaAgent").then((dest) => {
-          console.log('Succesfully downloaded ' + dest);
+          console.log('Successfully downloaded ' + dest);
           counter++;
           counter > 2 && resolve();
         }).catch(console.error);
